# perf(resolve): limits the #times external manifests are retrieved