## Description, Motivation and Context

- limits limits the #times external manifests are retrieved by checking for locality
- improves performance quite a bit, even with rules that need the contents of the manifests, like licenses and deprecations (roughly 1.5s for dependency-cruiser for a regular cruise with the same parameters as in PR #412, roughly 3s on react (regular cruise of 1189 modules/ 3692 dependencies taking ~12s instead of 15s))

## How Has This Been Tested?

- [x] automated non-regression tests
- [x] green ci

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [x] Refactor (non-breaking change which fixes an issue without changing functionality)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

- [x] :book:

  - My change doesn't require a documentation update, or ...
  - it _does_ and I have updated it

- [x] :balance_scale:
  - The contribution will be subject to [The MIT license](https://github.com/sverweij/dependency-cruiser/blob/develop/LICENSE), and I'm OK with that.
  - The contribution is my own original work.
  - I am ok with the stuff in [**CONTRIBUTING.md**](https://github.com/sverweij/dependency-cruiser/blob/develop/.github/CONTRIBUTING.md).